pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["Sergey Klimov", "Maxim Rukhlov", "Alex Erdyakov"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
Flask = "^2.0.1"
psycopg2-binary = "^2.9.1"
uvicorn = "^0.15.0"
Flask-SQLAlchemy = "^2.5.1"
flask-restx = "^0.5.0"
python-decouple = "^3.4"
Flask-Cors = "^3.0.10"
Flask-Bcrypt = "^0.7.1"
Flask-Migrate = "^3.1.0"

[tool.poetry.dev-dependencies]
pytest = "^6.2.4"
isort = "^5.9.3"
flake8 = "^3.9.2"
pytest-cov = "^2.12.1"
Flask-Admin = "^1.5.8"
ipython = "^7.26.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
